fixups for auth

+++ test/__init__.py
@@ -103,8 +103,7 @@ TEST_LOADBALANCER = bool(os.environ.get("TEST_LOADBALANCER"))
 TEST_SERVERLESS = bool(os.environ.get("TEST_SERVERLESS"))
 SINGLE_MONGOS_LB_URI = os.environ.get("SINGLE_MONGOS_LB_URI")
 MULTI_MONGOS_LB_URI = os.environ.get("MULTI_MONGOS_LB_URI")
-TEST_AUTH_AWS = os.environ.get("TEST_AUTH_AWS")
-TEST_AUTH_OIDC = os.environ.get("TEST_AUTH_OIDC")
+AUTH_MECH = os.environ.get("AUTH_MECH")
 
 if TEST_LOADBALANCER:
     res = parse_uri(SINGLE_MONGOS_LB_URI or "")
@@ -287,7 +286,6 @@ class ClientContext:
         self.is_data_lake = False
         self.load_balancer = TEST_LOADBALANCER
         self.serverless = TEST_SERVERLESS
-        self.no_admin = TEST_AUTH_AWS or TEST_AUTH_OIDC
         if self.load_balancer or self.serverless:
             self.default_client_options["loadBalanced"] = True
         if COMPRESSORS:
@@ -313,6 +311,8 @@ class ClientContext:
         opts = client_context.default_client_options.copy()
         opts.pop("server_api", None)  # Cannot be set from the URI
         opts_parts = []
+        if AUTH_MECH:
+            opts_parts.append(("authMechanism", AUTH_MECH))
         for opt, val in opts.items():
             strval = str(val)
             if isinstance(val, bool):
@@ -394,7 +394,7 @@ class ClientContext:
                 else:
                     self.auth_enabled = self._server_started_with_auth()
 
-            if self.auth_enabled and not self.no_admin:
+            if self.auth_enabled:
                 if not self.serverless:
                     # See if db_user already exists.
                     if not self._check_user_provided():
@@ -412,7 +412,7 @@ class ClientContext:
                 # May not have this if OperationFailure was raised earlier.
                 self.cmd_line = self.client.admin.command("getCmdLineOpts")
 
-            if self.serverless or self.no_admin:
+            if self.serverless:
                 self.server_status = {}
             else:
                 self.server_status = self.client.admin.command("serverStatus")
@@ -453,7 +453,7 @@ class ClientContext:
             self.w = len(hello.get("hosts", [])) or 1
             self.version = Version.from_client(self.client)
 
-            if self.serverless or self.no_admin:
+            if self.serverless:
                 self.server_parameters = {
                     "requireApiVersion": False,
                     "enableTestCommands": True,
@@ -479,7 +479,7 @@ class ClientContext:
             if self.is_mongos:
                 address = self.client.address
                 self.mongoses.append(address)
-                if not self.serverless and not self.no_admin:
+                if not self.serverless:
                     # Check for another mongos on the next port.
                     assert address is not None
                     next_address = address[0], address[1] + 1
@@ -1228,7 +1228,7 @@ def teardown():
         raise AssertionError("\n".join(garbage))
     c = client_context.client
     if c:
-        if not client_context.is_data_lake:
+        if not client_context.is_data_lake and not client_context:
             c.drop_database("pymongo-pooling-tests")
             c.drop_database("pymongo_test")
             c.drop_database("pymongo_test1")
